actions/requirements.txt
Pillow==9.0.0
pytesseract
pdf2image 
pdftotext
opencv-python-headless==4.1.2.30
easyocr
#poppler
#tesseract-ocr-setup-3.02.02 
google-auth==1.10.1 
prompt-toolkit==2.0.10 
questionary==1.4.0 
SQLAlchemy==1.3.12 
urllib3==1.25.7